Update issue 61 Startup parameters update was included to the software post-installation phase.
Installer detects which OS and Java version is running. Following changes are made in the osm2garmin.conf depending on the above:

- for 32bit Windows systems memory limit is set to -J-Xmx1300m which should work on most systems;
- for all other systems memory limit is set to -J-Xmx2000m;
- for all 64bit Java installations permGenSpace limits are set to -J-XX:PermSize=64M -J-XX:MaxPermSize=256m;
- for all non-windows systems CRLF line endings are replaced with LF only;
- jdkhome parameter is replaced by JAVAHOME installer is running on.

To change JVM for installer run, please run installer with --javahome <dir> parameter.

Uninstalling and reinstalling the software does not affect any program settings when dev dir is not deleted while uninstalling.